Fix sample_connected_subgraph being allowed to return tiny subgraphs
Previously, if the randomly sampled node from which to run BFS was on or
near a leaf, the resulting subgraph could be tiny. A such, the resulting
graph summary statistics would be unreliable.

To solve this, add arguments `min_nodes` and `max_attempts` to
sample_connected_subgraph(). The function will now subsample for
`max_attempts` until a subgraph with at least `min_nodes` is obtained.