fix(tuning): fix optuna_study trying to read empty Study db
When an Optuna study db is empty, checking if all trials are finished
evaluates to True while it shouldn't. optuna_study has short-circuit
logic to return the best hyperparameters in case a Study is already
finished, but that logic of course assumes a populated Study db. Doing
so with an empty db caused Optuna to try to read non-existent records,
causing a crash.